Refactor C tests to work with Zephyr
Update the c_test_framework header to further abstract the components
of the test framework to allow Zephyr to test the library using its
own tools (ztest and twister).

1. Abstract the main entry point for the test suite. Instead of the tests
   defining their own main() function, they implement fff_test_suite().
   A new main_c.c will run the common main function and link to the
   correct test suite.
2. Abstract printf so that Zephyr can use printk
3. Move c_test_framework.h into a framework/ directory. This allows Zephyr
   to avoid including this file in the path and instead provide its own
   version of the test framework.
