+++ test/ms_vc_fff_test/ms_vc_fff_test_global_c/ms_vc_fff_test_global_c.vcxproj.filters
@@ -21,12 +21,15 @@
     <ClCompile Include="..\..\global_fakes.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\main_c.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\global_fakes.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\c_test_framework.h">
+    <ClInclude Include="..\..\framework\c_test_framework.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="..\..\..\fff.h">
